Bug 1681096 - python3 - pylint --py3k - W1633: round built-in referenced (round-builtin) r=marionette-reviewers,perftest-reviewers,aki,gbrown,whimboo
Differential Revision: https://phabricator.services.mozilla.com/D99067

UltraBlame original commit: 1ed3facd004b4481026d72a50e8bf3e00df97eb2

+++ testing/talos/talos/run_tests.py
@@ -457,7 +457,8 @@ def make_comparison_result(base_and_reference_results):
         _index = 0
         for next_ref in comp_results[subtest_index]["ref_runs"]:
             diff = abs(next_ref - comp_results[subtest_index]["base_runs"][_index])
-            comp_results[subtest_index]["runs"].append(round(diff, 3))
+            
+            comp_results[subtest_index]["runs"].append(float(round(diff, 3)))
             _index += 1
